requirements.txt
streamlit==1.31.0
opencv-python==4.8.1.78
numpy==1.26.0
pillow==10.0.1
mediapipe==0.10.8
streamlit-image-comparison==0.0.4
scikit-image==0.22.0
tensorflow==2.15.0
matplotlib==3.8.0 